Refactor: 함수명 개선 및 코드 분리
