+++ RudyMark/RudyMark/Views/Common/Card/CardView.swift
@@ -14,9 +14,9 @@ struct CardView: View{
         ZStack {
             VStack(alignment: .leading,spacing: 20){
                 MainTitle(card: card)
-                ProgressViewSection(card: card)
+                TodayKcalMiniCard(card: card)
                 BloodCard(card: card)
-                MiniCardsSection(card: card)
+                NutrientsMiniCardContainer(card: card)
                 MealButtonsSection(card: card)
             }
             
@@ -30,47 +30,7 @@ struct CardView: View{
 }
 
 
-struct ProgressViewSection: View {
-    let card: CardData
-    
-    var body: some View {
-        if let progress = card.progress {
-            VStack(alignment: .center) {
-                HStack(alignment: .center, spacing: 10) {
-                    (
-                        Text("\(Int(progress))").bold().foregroundColor(.black).font(.title)
-                        + Text(" / \(Int(card.max ?? 2000)) kcal").foregroundColor(.gray).font(.title3)
-                    )
-                }
-                .frame(height: 80)
-                .frame(maxWidth: .infinity, alignment: .leading)
-                .padding(.leading, 16)
-                .background(Color(red: 0.94, green: 0.92, blue: 0.99))
-                .cornerRadius(20)
-            }
-        }
-    }
-}
 
-
-
-struct MiniCardsSection: View {
-    let card: CardData
-    
-    var body: some View {
-        if let miniCards = card.miniCards {
-            VStack{
-                ForEach(miniCards, id: \.title){ nutrient in
-                    HStack{
-                        MiniCardView(miniCard: nutrient,cardHeight: card.height)
-                    }
-                    .frame(maxWidth: .infinity,minHeight: card.miniCardsSize)
-                    .background(card.miniCardsColor)
-                }
-            }
-        }
-    }
-}
 
 struct MealButtonsSection: View {
     let card: CardData
@@ -171,61 +131,6 @@ struct MealButtonsSection: View {
     }
 }
 
-struct MiniCardView : View{
-    let miniCard: MiniCard
-    let cardHeight: CGFloat
-    
-    var body: some View{
-        VStack(alignment: .center){
-            if let progress = miniCard.progress, let max = miniCard.max, let barColor = miniCard.barColor {
-                HStack(alignment: .center,spacing:10) {
-                    Text(miniCard.title)
-                        .font(.subheadline)
-                        .foregroundColor(.deepGray)
-                    
-                    ZStack(alignment: .leading) {
-                        RoundedRectangle(cornerRadius: 20)
-                            .fill(Color.gray.opacity(0.3))
-                            .frame(width: 140, height: 16)
-                        RoundedRectangle(cornerRadius: 20)
-                            .fill(barColor)
-                            .frame(width: min(CGFloat(progress / max) * 140, 140), height: 16)
-                    }
-                    
-                    (
-                        Text("\(Int(progress))").bold().foregroundColor(.black)
-                        + Text(" / \(Int(max)) g").foregroundColor(.black)
-                    )
-                }
-                .frame(maxWidth: .infinity, alignment: .center)
-            }
-        }
-    }
-}
 
 
 
-struct ArcShape: Shape {
-    var progress: Double
-    
-    func path(in rect: CGRect) -> Path {
-        var path = Path()
-        let radius = min(rect.width, rect.height) / 1.5
-        let center = CGPoint(x: rect.midX, y: rect.midY)
-        let startAngle = Angle.degrees(0)
-        let endAngle = Angle.degrees(180 * progress)
-        
-        path.addArc(center: center,
-                    radius: radius,
-                    startAngle: startAngle,
-                    endAngle: endAngle,
-                    clockwise: false)
-        path = path.strokedPath(StrokeStyle(lineWidth: 10, lineCap: .round))
-        return path
-    }
-    
-    var animatableData: Double {
-        get { progress }
-        set { progress = newValue }
-    }
-}
